+++ browserstep/common.py
@@ -9,6 +9,26 @@ from behave import *
 from selenium import webdriver
 from selenium.webdriver.support.ui import Select
 
+from selenium.webdriver.common.action_chains import ActionChains
+
+
+@step('I hover over "{container_selector}"')
+def step_impl(context, container_selector):
+    element_to_hover_over = context.browser.find_element_by_css_selector(container_selector)
+    hover = ActionChains(context.browser).move_to_element(element_to_hover_over)
+    hover.perform()
+    time.sleep(0.7)
+
+
+@step('I follow the "{text}" link in "{container_selector}"')
+def step_impl(context, text, container_selector):
+    container = context.browser.find_element_by_css_selector(container_selector)
+    elements = container.find_elements_by_link_text(text)
+    if not elements:
+        elements = container.find_elements_by_xpath("//img[contains(@alt,'{}')]".format(text))
+    assert len(elements) == 1, "Expected 1 matching link, not {}".format(len(elements))
+    elements[0].click()
+
 
 @step('I navigate to /{path}')
 def step_impl(context, path):